// src/api/model/foodDTO.ts
export interface FoodDTO {
  id?: number;
  version?: number;
  createdDate?: Date;
  lastModifiedDate?: Date;
  brandName?: string;
  description?: string;
  calories?: number;
  proteins?: number;
  totalCarbohydrates?: number;
  fibers?: number;
  sugars?: number;
  totalFats?: number;
  saturatedFats?: number;
  polyUnsaturatedFats?: number;
  monoUnsaturatedFats?: number;
  transFats?: number;
  cholesterol?: number;
  sodium?: number;
  potassium?: number;
  calcium?: number;
  iron?: number;
  vitaminA?: number;
  vitaminB1?: number;
  vitaminB2?: number;
  vitaminB3?: number;
  vitaminB5?: number;
  vitaminB6?: number;
  vitaminB7?: number;
  vitaminB9?: number;
  vitaminB12?: number;
  vitaminC?: number;
  vitaminD?: number;
  vitaminE?: number;
  vitaminK?: number;
  omega3?: number;
  omega6?: number;
  omega9?: number;
}